[418] EntityPart to core pkg ; new content methods
- Moves `EntityPart` from `ext` to `core` package
- Adds new `content` methods to `EntityPart.Builder` that use MBWs
- Cleanup of references to old `Part` name updated to `EntityPart`
