+++ examples/src/main/java/jaxrs/examples/multipart/MultipartClient.java
@@ -21,10 +21,11 @@ import jakarta.ws.rs.client.Client;
 import jakarta.ws.rs.client.ClientBuilder;
 import jakarta.ws.rs.client.Entity;
 import jakarta.ws.rs.client.WebTarget;
+import jakarta.ws.rs.core.EntityPart;
 import jakarta.ws.rs.core.GenericType;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
-import jakarta.ws.rs.ext.EntityPart;
+
 
 public class MultipartClient {
     Logger LOG = Logger.getLogger(MultipartClient.class.getName());
