+++ jaxrs-api/src/test/java/jakarta/ws/rs/core/RuntimeDelegateStub.java
@@ -22,11 +22,12 @@ import jakarta.ws.rs.SeBootstrap;
 import jakarta.ws.rs.SeBootstrap.Configuration;
 import jakarta.ws.rs.SeBootstrap.Instance;
 import jakarta.ws.rs.core.Application;
+import jakarta.ws.rs.core.EntityPart;
 import jakarta.ws.rs.core.UriBuilder;
 import jakarta.ws.rs.core.Link.Builder;
 import jakarta.ws.rs.core.Response.ResponseBuilder;
 import jakarta.ws.rs.core.Variant.VariantListBuilder;
-import jakarta.ws.rs.ext.EntityPart;
+
 import jakarta.ws.rs.ext.RuntimeDelegate;
 
 public class RuntimeDelegateStub extends RuntimeDelegate {
