+++ jaxrs-spec/src/main/asciidoc/chapters/resources/_declaring_method_capabilities.adoc
@@ -137,7 +137,7 @@ The `multipart/form-data` media type enables requests to send multiple
 entities (parts) as a single entity. Each part contains its own set of
 headers, media type and content. There are a few different ways to
 consume these requests in a resource method. First, a resource method
-may contain an entity parameter of type, `List<Part>`. For example:
+may contain an entity parameter of type, `List<EntityPart>`. For example:
 
 [source,java]
 ----
@@ -146,8 +146,8 @@ public class WidgetsResource3 {
 
     @POST
     @Consumes(MediaType.MULTIPART_FORM_DATA)
-    public Response postWidget(List<Part> parts) {
-        for (Part part : parts) {
+    public Response postWidget(List<EntityPart> parts) {
+        for (EntityPart part : parts) {
             String name = part.getName();
             Optional<String> fileName = part.getFileName();
             InputStream is = part.getContent();
@@ -163,8 +163,8 @@ public class WidgetsResource3 {
 
 Another approach is to use `@FormParam` parameters where the value in the
 annotation corresponds to the name of the part. The parameter type may be
-a `jakarta.ws.rs.ext.Part`, a `java.io.InputStream`, a `String`, or any type
-that can be converted to using a registered `MessageBodyReader` or
+a `jakarta.ws.rs.core.EntityPart`, a `java.io.InputStream`, a `String`, or
+any type that can be converted to using a registered `MessageBodyReader` or
 `ParamConverter`. Here is
 an example:
 
@@ -177,16 +177,17 @@ public class WidgetsResource4 {
     @Consumes(MediaType.MULTIPART_FORM_DATA)
     public Response postWidget(@FormParam("part1Name") String part1, 
                                @FormParam("part2Name") InputStream part2,
-                               @FormParam("part3Name") Part part3,
+                               @FormParam("part3Name") EntityPart part3,
                                @FormParam("part4Name") Widget part4) {...}
 }
 ----
 
-The only way to access the headers for a particular part is to use the `Part`
-type. The `InputStream`, `String`, and additional parameter types will only
-provide the content of the part. Note that parts of a multipart entity can be
-quite large, so care should be taken when using `String` parameter types as
-that will load the entire content of the part into the Java heap.
+The only way to access the headers for a particular part is to use the
+`EntityPart` type. The `InputStream`, `String`, and additional parameter
+types will only provide the content of the part. Note that parts of a
+multipart entity can be quite large, so care should be taken when using
+`String` parameter types as that will load the entire content of the part
+into the Java heap.
 
 When converting the part's content, the implementation MUST use the encoding
 charset specified in the part's `Content-Type` header, or `UTF-8` if the
