+++ jaxrs-spec/src/main/asciidoc/chapters/resources/_resource_method.adoc
@@ -47,7 +47,7 @@ property values, see <<resource_field>>. Exceptions thrown during
 construction of `@FormParam` annotated parameter values are treated the
 same as if the parameter were annotated with `@HeaderParam`.
 * For resource methods that consume `multipart/form-data` media types,
-the `@FormParam` annotation may be used on `jakarta.ws.rs.ext.Part`
+the `@FormParam` annotation may be used on `jakarta.ws.rs.core.EntityPart`
 parameters. See <<consuming_multipart_formdata>> for more details.
 
 [[entity_parameters]]
